requirements.txt
dash==2.9.3
dash_bootstrap_components==1.2.0
dash_bootstrap_templates==1.0.8
dash_core_components==2.0.0
dash_html_components==2.0.0
dash_table==5.0.0
numpy==1.21.6
pandas==1.3.5
plotly==5.9.0
psycopg2_binary==2.9.6
waitress==1.2.0b2
Werkzeug==2.0.3
